AddAction, EditAction, Transaction double changed to BigDecimal

+++ src/main/java/ui/TransactionsTable.java
@@ -5,6 +5,7 @@ import model.Category;
 import model.Transaction;
 import model.TransactionType;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -13,7 +14,7 @@ public class TransactionsTable extends AbstractEntityTableModel<Transaction> {
 
     private static final List<Column<?, Transaction>> COLUMNS = List.of(
             Column.readOnly("Name", String.class, Transaction::getName),
-            Column.readOnly("Amount", Double.class, Transaction::getAmount),
+            Column.readOnly("Amount", BigDecimal.class, Transaction::getAmount),
             Column.readOnly("Type", TransactionType.class, Transaction::getType),
             Column.readOnly("Category", Category.class, Transaction::getCategory),
             Column.readOnly("Created", Date.class, Transaction::getDate),
